// src/types.ts
export interface Voter {
  EpicNumber: string;
  ACNo: number;
  PartNo: number;
  SerialNo: number;
  ElectorsName: string;
  ElectorNameHindi: string;
  ElectorGender: string;
  Age: number;
  DOB: string;
  RelativeName: string;
  RelativeNameHindi: string;
  Relativetype: string;
  AdharNumber: string;
  MobileNumber: string;
}